+++ maya_umbrella_launcher/launcher_cmd.py
@@ -2,9 +2,9 @@ import os
 import argparse
 
 
-import translator as tr
-from core import (get_python_path_env, launch_maya, install_to_maya, set_plugin_folder,
-                  download_plugin, get_maya_app_path, uninstall_from_maya, get_plugin_folder)
+import maya_umbrella_launcher.translator as tr
+from maya_umbrella_launcher.filesystem import MayaSystem
+from maya_umbrella_launcher.core import PluginManager, PluginInstaller, UserSetting
 
 
 def main():
@@ -20,32 +20,32 @@ def main():
 
     if args.path:
         if os.path.isdir(args.path):
-            set_plugin_folder(args.path)
+            UserSetting.set('plugin_folder', args.path)
         else:
             print(tr.path_not_exists.text.format(args.path))
             return
     elif args.path == '':
-        print(get_plugin_folder())
+        print(UserSetting.get('plugin_folder'))
 
     if args.download:
-        if not download_plugin(overwrite=True):
+        if not PluginManager.download_plugin(overwrite=True):
             return
 
     if args.start:
-        envs = get_python_path_env()
-        maya_app_path = get_maya_app_path(maya_version=args.start)
-        launch_maya(maya_path=maya_app_path,
-                    envs=envs
-                    )
+        envs = PluginManager.get_python_path_env()
+        maya_app_path = MayaSystem.get_maya_app_path(maya_version=args.start)
+        MayaSystem.launch_maya(maya_path=maya_app_path,
+                               envs=envs
+                               )
         return
 
     if args.install:
-        install_to_maya(maya_version=args.install)
+        PluginInstaller.install(maya_version=args.install)
         print(tr.run_install.text)
         return
 
     if args.uninstall:
-        uninstall_from_maya(maya_version=args.uninstall)
+        PluginInstaller.uninstall(maya_version=args.uninstall)
         print(tr.run_uninstall.text)
         return
